Removing unused `getSimpleName`
This method is not used throughout the project.  Instead of commenting
out code, I'm removing that code.

As a matter of reading code, it's hard to determine if commented out
code is intended to be kept around.

I strongly prefer to remove code and leave a meaningful comment to explain.