Refactoring away removeSpacesObjectProperities()
Prior to this commit, we had a function that would iterate over an
object and remove spaces from the named property's values.

Instead, this method now adds configuration options to the
`normalizeObjectProperties()`. In this way, we continue to leverage the
common functions in `./ingest/units/normalizer.ts`

This also includes a minor bit of clean-up to consolidate two function
calls into a single parameterized function call.