Extracting a getAllSimplifiedPersons function
This function was common throughout the updated files.  Instead of
having variance, I want to normalize these functions.

One thing that I'm dis-satisfied with is that there are lots of
ApolloClient declarations.  To resolve this for the extraction of a
common method, I went with dependency injection.

I don't know the best way to proceed with this, but for now this change
should be logically equivalent to the code state prior to this commit.

Also, I extracted what appeared to be another "normalize the data" kind
of method associated with the simplified person.

NOTE: There remains a `getAllSimplifiedPersons()` that is different in
its inner logic.  I'm uncertain if that is a refinement that should
propogate to other locations or if it is a specific implementation.