Removing unused methods and commented code
The lsp-mode compiler says that all of the removed methods are not
called.  I'm also removing the commented out methods as they add to the
chatter.  As someone new to the project, it's hard to discern what the
purpose of the file is when there is commented out code AND uncalled
functions.

If this is not something to remove, that's fine, but I believe the file
needs revisiting.