Expanding NormedPerson interface to class
This commit moves load from CSV logic closer to the data structure
interface.

And while it doesn't check off all of the TODO items for #137, it helps
tidy up the code base.

I'm going to be reviewing the details of those checkbox items, as I'm
unclear the purpose (e.g. For example, writing the data structure to CSV)